+++ src/components/pubsub/SubscriptionCreate.tsx
@@ -102,7 +102,8 @@ function SubscriptionCreate({
     <>
       <Box
         component="form"
-        name="subscition_create"
+        id="subscription_create"
+        name="subscription_create"
         noValidate
         autoComplete="off"
         className="flex flex-col gap-2"
@@ -113,7 +114,22 @@ function SubscriptionCreate({
           <Controller
             name="name"
             control={control}
-            rules={{ required: true }}
+            rules={{ 
+              validate: {
+                checkFormat: (name: string) => {
+                    const regex = /^[a-zA-Z]{1}[a-zA-Z0-9\-_%+~]{2,254}$/i;
+                    if (!regex.test(name)) {
+                      return 'Subscription name format is not correct';
+                    }
+                },          
+                checkName: (name: string) => {
+                    if (name.toLowerCase().includes('goog')) {
+                      return 'Subscription name cannot contain "goog"';
+                    }
+                }
+              },
+              required: true
+            }}
             render={({ field }) => (
               <TextField
                 {...field}
@@ -177,7 +193,7 @@ function SubscriptionCreate({
             <Alert severity="error">{SubmitError}</Alert>
           )}
           {IsCreated && (
-            <Alert severity="success">Subscription is created</Alert>
+            <Alert severity="success">Subscription created</Alert>
           )}
         </Stack>