+++ src/components/pubsub/TopicCreate.tsx
@@ -108,6 +108,7 @@ function TopicCreate({
     <>
       <Box
         component="form"
+        id="topic_create"
         name="topic_create"
         noValidate
         autoComplete="off"
@@ -119,7 +120,22 @@ function TopicCreate({
           <Controller
             name="name"
             control={control}
-            rules={{ required: true }}
+            rules={{ 
+              validate: {
+                checkFormat: (name: string) => {
+                    const regex = /^[a-zA-Z]{1}[a-zA-Z0-9\-_%+~]{2,254}$/i;
+                    if (!regex.test(name)) {
+                      return 'Topic name format is not correct';
+                    }
+                },          
+                checkName: (name: string) => {
+                    if (name.toLowerCase().includes('goog')) {
+                      return 'Topic name cannot contain "goog"';
+                    }
+                }
+              },
+              required: true
+            }}
             render={({ field }) => (
               <Tooltip title="Topic name" placement="top-start">
                 <TextField
@@ -130,6 +146,7 @@ function TopicCreate({
                   size="small"
                   variant="filled"
                   error={errors.name ? true : false}
+                  helperText={errors.name?.message}
                 />
               </Tooltip>
             )}
@@ -170,7 +187,7 @@ function TopicCreate({
           {SubmitError != undefined && (
             <Alert severity="error">{SubmitError}</Alert>
           )}
-          {IsCreated && <Alert severity="success">Topic is created</Alert>}
+          {IsCreated && <Alert severity="success">Topic created</Alert>}
         </Stack>
 
         <Collapse in={isAdvanced}>
@@ -225,7 +242,7 @@ function TopicCreate({
                     <Tooltip title="Schema" placement="top-start">
                       <Select
                         {...field}
-                        id="scheme"
+                        id="schema-name"
                         labelId="schema-name-label"
                         size="small"
                         variant="filled"