+++ vite.config.ts
@@ -15,6 +15,9 @@ export default defineConfig(async () => ({
     port: 1420,
     strictPort: true,
   },
+  preview: {
+    port: 8088,
+  },
   // 3. to make use of `TAURI_DEBUG` and other env variables
   // https://tauri.studio/v1/api/config#buildconfig.beforedevcommand
   envPrefix: [